Struct sucds::bit_vectors::darray::DArray

pub struct DArray { /* private fields */ }

Constant-time select data structure over integer sets with the dense array technique.

Memory complexity

$u + o(u)$ bits for a bit vector with $u$ bits.

Notes

In the default configuration, this data structure supports only Self::select1(). If rank queries are needed, Self::enable_rank() and Self::enable_select0() must be set up.

Examples

use sucds::bit_vectors::{DArray, Access, Rank, Select};

let da = DArray::from_bits([true, false, false, true])
    .enable_rank()
    .enable_select0();

assert_eq!(da.len(), 4);
assert_eq!(da.access(1), Some(false));

assert_eq!(da.rank1(1), Some(1));
assert_eq!(da.rank0(1), Some(0));

assert_eq!(da.select1(1), Some(3));
assert_eq!(da.select0(0), Some(1));

Credits

This is a yet another Rust port of succinct::darray.

References

D. Okanohara, and K. Sadakane, “Practical Entropy-Compressed Rank/Select Dictionary,” In ALENEX, 2007.

Implementations

impl DArray

pub fn from_bits<I>(bits: I) -> Self
where I: IntoIterator<Item = bool>,

Creates a new instance from input bit stream bits.

Arguments

bits: Bit stream.

pub fn enable_rank(self) -> Self

Builds an index to enable rank queries.

pub fn enable_select0(self) -> Self

Builds an index to enable select0.

pub const fn has_rank(&self) -> bool

Checks if Self::enable_rank() is set.

pub const fn has_select0(&self) -> bool

Checks if Self::enable_select0() is set.

pub const fn bit_vector(&self) -> &BitVector

Returns the reference of the internal bit vector.

pub const fn s1_index(&self) -> &DArrayIndex

Returns the reference of the internal select1 index.

pub const fn s0_index(&self) -> Option<&DArrayIndex>

Returns the reference of the internal select0 index.

pub const fn r9_index(&self) -> Option<&Rank9SelIndex>

Returns the reference of the internal rank index.

pub const fn len(&self) -> usize

Returns the number of bits stored.

pub const fn is_empty(&self) -> bool

Checks if the vector is empty.