Struct sucds::bit_vectors::rank9sel::Rank9Sel

pub struct Rank9Sel { /* private fields */ }

Rank/select data structure over bit vectors with Vigna’s rank9 and hinted selection techniques.

This builds rank/select indices on BitVector taking

25% overhead of space for the rank index, and

3% overhead of space for the select index (together with the rank’s overhead).

Notes

In the default configuration, it does not build the select index for faster queires. To accelerate the queries, set Self::select1_hints() and Self::select0_hints().

Examples

use sucds::bit_vectors::{Rank9Sel, Access, Rank, Select};

let bv = Rank9Sel::from_bits([true, false, false, true])
    .select1_hints()
    .select0_hints();

assert_eq!(bv.len(), 4);
assert_eq!(bv.access(1), Some(false));

assert_eq!(bv.rank1(1), Some(1));
assert_eq!(bv.rank0(1), Some(0));

assert_eq!(bv.select1(1), Some(3));
assert_eq!(bv.select0(0), Some(1));

Credits

This is a yet another Rust port of succinct::rs_bit_vector.

References

S. Vigna, “Broadword implementation of rank/select queries,” In WEA, 2008.

Implementations

impl Rank9Sel

pub fn new(bv: BitVector) -> Self

Creates a new vector from input bit vector bv.

pub fn select1_hints(self) -> Self

Builds an index for faster select1.

pub fn select0_hints(self) -> Self

Builds an index for faster select0.

pub fn from_bits<I>(bits: I) -> Self
where I: IntoIterator<Item = bool>,

Creates a new vector from input bit stream bits.

Arguments

bits: Bit stream.

pub const fn bit_vector(&self) -> &BitVector

Returns the reference of the internal bit vector.

pub const fn rs_index(&self) -> &Rank9SelIndex

Returns the reference of the internal rank/select index.

pub const fn len(&self) -> usize

Returns the number of bits stored.

pub const fn is_empty(&self) -> bool

Checks if the vector is empty.