Struct sucds::bit_vectors::sarray::SArray

pub struct SArray { /* private fields */ }

Rank/Select data structure over very sparse bit vectors, which is a specialized version of EliasFano for bit vectors.

Memory complexity

$n \lceil \lg \frac{u}{n} \rceil + 2n + o(n)$ bits for a bit vector with $u$ bits and $n$ set bits.

Notes

This data structure does not support select0.

Examples

use sucds::bit_vectors::{SArray, Access, Rank, Select};

let sa = SArray::from_bits([true, false, false, true]).enable_rank();

assert_eq!(sa.len(), 4);
assert_eq!(sa.access(1), Some(false));

assert_eq!(sa.rank1(1), Some(1));
assert_eq!(sa.rank0(1), Some(0));
assert_eq!(sa.select1(1), Some(3));

References

D. Okanohara, and K. Sadakane, “Practical Entropy-Compressed Rank/Select Dictionary,” In ALENEX, 2007.

Implementations

impl SArray

pub fn from_bits<I>(bits: I) -> Self
where I: IntoIterator<Item = bool>,

Creates a new instance from input bit stream bits.

Arguments

bits: Bit stream.

pub fn enable_rank(self) -> Self

Builds an index to enable rank, predecessor, and successor queries.

pub const fn has_rank(&self) -> bool

Checks if Self::enable_rank() is set.

pub fn predecessor1(&self, pos: usize) -> Option<usize>

Returns the largest bit position pred such that pred <= pos and the pred-th bit is set, or None if not found or self.len() <= pos.

Arguments

pos: Bit position.

Complexity

$O(\lg \frac{u}{n})$

Panics

It panics if the index is not built by Self::enable_rank().

Examples

use sucds::bit_vectors::SArray;

let sa = SArray::from_bits([false, true, false, true]).enable_rank();

assert_eq!(sa.predecessor1(3), Some(3));
assert_eq!(sa.predecessor1(2), Some(1));
assert_eq!(sa.predecessor1(1), Some(1));
assert_eq!(sa.predecessor1(0), None);

pub fn successor1(&self, pos: usize) -> Option<usize>

Returns the smallest bit position succ such that succ >= pos and the succ-th bit is set, or None if not found or self.len() <= pos.

Arguments

pos: Bit position.

Complexity

$O(\lg \frac{u}{n})$

Panics

It panics if the index is not built by Self::enable_rank().

Examples

use sucds::bit_vectors::SArray;

let sa = SArray::from_bits([true, false, true, false]).enable_rank();

assert_eq!(sa.successor1(0), Some(0));
assert_eq!(sa.successor1(1), Some(2));
assert_eq!(sa.successor1(2), Some(2));
assert_eq!(sa.successor1(3), None);

pub const fn len(&self) -> usize

Returns the number of bits stored.

pub const fn is_empty(&self) -> bool

Checks if the vector is empty.