Trait sucds::bit_vectors::Rank

pub trait Rank {
    // Required methods
    fn rank1(&self, x: usize) -> Option<usize>;
    fn rank0(&self, x: usize) -> Option<usize>;
}

Interface for rank queries on bit vectors.

Let $S \subseteq \{ 0,1,\dots,u-1 \}$ be a set of positions at which bits are set in a bit vector of length $u$.

Required Methods

fn rank1(&self, x: usize) -> Option<usize>

Returns the cardinality of $\{ x \in S \mid x < i \}$, or None if $u < x$.

fn rank0(&self, x: usize) -> Option<usize>

Returns the cardinality of $\{ x \not\in S \mid 0 \leq x < i \}$, or None if $u < x$.