Struct sucds::mii_sequences::elias_fano::EliasFanoBuilder

pub struct EliasFanoBuilder { /* private fields */ }

Builder for EliasFano.

Examples

use sucds::mii_sequences::EliasFanoBuilder;

let mut efb = EliasFanoBuilder::new(8, 5)?;

assert_eq!(efb.universe(), 8);
assert_eq!(efb.num_vals(), 5);

efb.push(1)?;
efb.push(3)?;
efb.extend([3, 5, 7])?;

let ef = efb.build();
assert_eq!(ef.len(), 5);
assert_eq!(ef.universe(), 8);

Implementations

impl EliasFanoBuilder

pub fn new(universe: usize, num_vals: usize) -> Result<Self>

Creates a new builder.

Arguments

universe: The (exclusive) upper bound of integers to be stored, i.e., an integer in [0..universe - 1].

num_vals: The number of integers that will be pushed (> 0).

Errors

An error is returned if num_vals == 0.

pub fn push(&mut self, val: usize) -> Result<()>

Pushes integer val at the end.

Arguments

val: Pushed integer that must be no less than the last one.

Errors

An error is returned if

val is less than the last one,

val is no less than Self::universe(), or

the number of stored integers becomes no less than Self::num_vals().

pub fn extend<I>(&mut self, vals: I) -> Result<()>
where I: IntoIterator<Item = usize>,

Appends integers at the end.

Arguments

vals: Pushed integers that are monotone increasing.

Errors

An error is returned if

vals is not monotone increasing (also compared to the current last value),

values in vals is no less than Self::universe(), or

the number of stored integers becomes no less than Self::num_vals().

pub fn build(self) -> EliasFano

Builds EliasFano from the pushed integers.

pub const fn universe(&self) -> usize

Returns the universe, i.e., the (exclusive) upper bound of possible integers.

pub const fn num_vals(&self) -> usize

Returns the number of integers that can be stored.